Fixed bug with updating candidate list when creating, removing, or updating a tab.
The bug was that the list would only be updated correctly if we were on
the tab holding the list; otherwise, on non-active tabs, the list would
not be updated.

Furthermore, if a highlighted tab was removed, then the first tab in the
candidate list would be highlighted next. I did not consider this to be
good behavior and changed it so that it would highlight the active tab
instead.